[Refactor] remove HFT sw example [sw] start implementing CMAC and AXI Stream FIFO driver
