--- sw/SoC/examples/smart_nic/ld/user.ld
@@ -1,9 +1,0 @@
-/* 
-    *** User-defined linker script ***
-
-    If you want to extend the UninaSoC.ld script, place here your code.
-    If you want to redefine your own linker script, remove the UninaSoC include.
-
-*/
-
-INCLUDE ../../common/UninaSoC.ld
